+++ Backend/Repositories/IFavorDishRepository.cs
@@ -4,7 +4,7 @@ namespace Backend.Repositories;
 
 public interface IFavorDishRepository
 {
-    Task<List<FavoriteDish>> GetList(int idUser);
+    Task<List<FavoriteDish>> GetList(long idUser);
     
     Task Add(FavoriteDish favoriteDish);
 }
